Segundo commit

+++ src/main/java/com/samuel/oficina/service/CargoService.java
@@ -6,14 +6,17 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.samuel.oficina.Exception.BusinessException;
 import com.samuel.oficina.domain.Cargo;
 import com.samuel.oficina.repository.CargoRepository;
+
 
 @Service
 public class CargoService {
 
 	@Autowired
 	private CargoRepository cargoRepository;
+
 	
 	public List<Cargo> listarTodosOsCargos() {
 		List<Cargo> cargos = cargoRepository.findAll();
@@ -24,11 +27,17 @@ public class CargoService {
 	}
 	
 	public Cargo inserirCargo(Cargo cargo) {
+		if (cargo.getNome().equals("")) {
+			throw new BusinessException("O nome é obrigatório!");
+		}
 		return cargoRepository.save(cargo);
 	}
 	
 	public Cargo excluirCargo(short id) {
 		Optional<Cargo> cargoEncontrado = cargoRepository.findById(id);
+		if (cargoEncontrado.isEmpty()) {
+			throw new RuntimeException("Cargo não encontrado.");
+		}
 		cargoRepository.delete(cargoEncontrado.get());
 		return cargoEncontrado.get();
 	}
